# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gym Project API
  version: 1.0.0
  description: JSON REST API interacting with the gym project backend
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: auth_key
      description: Authentication key obtained from /auth/login
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          example: johndoe
        password:
          type: string
          minLength: 6
          example: password123
    LoginResponse:
      type: object
      properties:
        auth_key:
          type: string
          example: a3f9c8b1e7d44e3c9f5b4d2a1c8e6f9a
        userId:
          type: integer
          example: 1
        role:
          type: string
          example: admin
        name:
          type: string
          example: John Doe
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - first_name
        - last_name
        - role
        - phone
      properties:
        username:
          type: string
          minLength: 3
          example: johndoe
        password:
          type: string
          minLength: 6
          example: Password@123
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        role:
          type: string
          enum: [admin, member, trainer]
          example: member
        phone:
          type: string
          example: "9876543210"
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        userId:
          type: integer
          example: 10
...
paths:
...
  /auth/login:
    post:
      summary: Login user and generate authentication key
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    delete:
      summary: Logout user
      tags:
        - Authentication
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== POSTS ROUTES ====================
...
security:
  - ApiKeyAuth: []